Read list of possible pairs from any config
Previously, the currently active pairs were set in `.git/config` while
the list of possible pairs was in `~/.gitconfig`. This allows the list
of possible pairs to be shared across multiple repos while still having
per-repo configuration for active pairs.

However, if you want to pre-populate a list of possible pairs, it is
handy to be able to set them system-wide, i.e. in /etc/gitconfig

This commit changes all the *reads* of possible pairs to read from
system, global or local, in order to allow system administrators to
pre-populate a list in /etc/gitconfig. But *writes* of possible pairs
(i.e. `git pair add`) still go to `~/.gitconfig` so that individual
users can add new pairs without needing root access.